requirements.txt
hydra-core==1.1.1
numpy==1.20.1
omegaconf==2.1.1
opencv_python_headless==4.5.3.56
PyYAML==6.0
scipy==1.4.1
tqdm==4.59.0
